requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.3
Django==4.1.5
django-auditlog==2.2.1
django-cors-headers==3.13.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.1
packaging==22.0
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.4.0
